remove specific commands as they are not really necessary
