+++ src/main/java/com/wire/backups/exports/commands/ExporterProducer.java
@@ -1,0 +1,12 @@
+package com.wire.backups.exports.commands;
+
+import com.wire.backups.exports.exporters.ExportConfiguration;
+import com.wire.backups.exports.exporters.Exporter;
+
+import javax.ws.rs.client.Client;
+
+@FunctionalInterface
+public interface ExporterProducer {
+    Exporter build(ExportConfiguration one, Client two);
+}
+
