+++ src/main/java/com/wire/backups/exports/exporters/Exporter.java
@@ -1,0 +1,8 @@
+package com.wire.backups.exports.exporters;
+
+public interface Exporter {
+    /**
+     * Executes exporter.
+     */
+    void run() throws Exception;
+}
